ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
openai==1.3.7
cohere==4.37
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
structlog==23.2.0
